add missing condition

+++ vaadin-grid-flow-parent/vaadin-grid-flow/src/main/java/com/vaadin/flow/data/provider/hierarchy2/HierarchicalDataCommunicator.java
@@ -147,10 +147,12 @@ public class HierarchicalDataCommunicator<T> extends DataCommunicator<T> {
      */
     @Override
     public void reset() {
-        getKeyMapper().removeAll();
-        dataGenerator.destroyAllData();
+        if (rootCache != null) {
+            rootCache = null;
+            getKeyMapper().removeAll();
+            dataGenerator.destroyAllData();
+        }
 
-        rootCache = null;
         requestFlush();
     }
